requirements.txt
numpy==2.3.1
QuantLib-Python==1.38
matplotlib==3.10.3
pandas ==2.3.2